requirements_ndb.txt
# for google cloud platform

bottle

pyjwt
social-auth-core

#for email: https://github.com/gsuitedevs/python-samples/blob/master/gmail/quickstart/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

grpcio
google-cloud-ndb
google-cloud-storage

# # DEBUG
# google-python-cloud-debugger
# # needs logging permission
# google-cloud-logging

numpy
matplotlib
lxml
sympy

#content
chcko-r
